+++ neo/renderer/tr_subview.cpp
@@ -524,7 +524,7 @@ bool	R_GenerateSurfaceSubview( drawSurf_t *drawSurf ) {
 	} // DG end
 
 	// see what kind of subview we are making
-	if ( shader->GetSort() != SS_SUBVIEW ) {
+	if ( shader->GetSort() != (float)SS_SUBVIEW ) {
 		for ( int i = 0 ; i < shader->GetNumStages() ; i++ ) {
 			const shaderStage_t	*stage = shader->GetStage( i );
 			switch ( stage->texture.dynamic ) {
